+++ src/mcpd.h
@@ -42,7 +42,7 @@
 #include "MCPTransportBLE.h"
 #endif
 
-#define MCPD_VERSION "0.27.4"
+#define MCPD_VERSION "0.27.5"
 #define MCPD_MCP_PROTOCOL_VERSION "2025-03-26"
 
 namespace mcpd {