+++ src/tools/MCPESPNowTool.h
@@ -69,7 +69,7 @@ static bool _parseMac(const char* str, uint8_t* mac) {
                   &mac[0], &mac[1], &mac[2], &mac[3], &mac[4], &mac[5]) == 6;
 }
 
-static ESPNowPeerStats* _findPeerStats(const uint8_t* mac) {
+[[maybe_unused]] static ESPNowPeerStats* _findPeerStats(const uint8_t* mac) {
     for (auto& ps : _espnowPeerStats) {
         if (memcmp(ps.mac, mac, 6) == 0) return &ps;
     }